backend/requirements.txt
# Core Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & AI
langchain==0.1.0
langchain-groq==0.0.1
langchain-openai==0.0.5
langchain-community==0.0.10

# Vector Store & Embeddings
chromadb==1.3.4
sentence-transformers==5.1.2

# HTTP & Async
# Pin httpx<0.28 due to breaking changes in TestClient API
httpx>=0.25.2,<0.28
aiofiles==23.2.1

# Security & Configuration
cryptography==41.0.7
python-dotenv==1.0.0
python-multipart==0.0.6

# Retry Logic
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging
structlog==23.2.0

# Utilities
python-json-logger==2.0.7
langgraph>=1.0.3

# Export & Document Generation
weasyprint>=66.0
markdown>=3.10
jinja2>=3.1.6
